refactor(shapes): eliminates the need for standalone shape coponents
Shape components are now compiled and exported from BasicShapes.js
