--- src/components/Circle.js
@@ -1,9 +1,0 @@
-import docs from '../docs/docs';
-import createSimpleComponent from '../utils/createSimpleComponent';
-
-export const Circle = (props) => {
-  const doc = docs.circle;
-  return createSimpleComponent(doc, props);
-};
-
-Circle.displayName = 'Circle';
