--- src/components/Polygram.js
@@ -1,7 +1,0 @@
-import docs from '../docs/docs';
-import createSimpleComponent from '../utils/createSimpleComponent';
-
-export const Polygram = (props) => {
-  const doc = docs.polygram;
-  return createSimpleComponent(doc, props);
-};
