--- src/components/RadialLines.js
@@ -1,7 +1,0 @@
-import docs from '../docs/docs';
-import createSimpleComponent from '../utils/createSimpleComponent';
-
-export const RadialLines = (props) => {
-  const doc = docs.radialLines;
-  return createSimpleComponent(doc, props);
-};
