--- src/components/Sector.js
@@ -1,7 +1,0 @@
-import docs from '../docs/docs';
-import createSimpleComponent from '../utils/createSimpleComponent';
-
-export const Sector = (props) => {
-  const doc = docs.sector;
-  return createSimpleComponent(doc, props);
-};
