--- src/components/Triangle.js
@@ -1,7 +1,0 @@
-import docs from '../docs/docs';
-import createSimpleComponent from '../utils/createSimpleComponent';
-
-export const Triangle = (props) => {
-  const doc = docs.triangle;
-  return createSimpleComponent(doc, props);
-};
